# Small optimizations to base64decode