At an airport so just looking for small stuff :)

- Don't copy input string to bytes since indexing by number indexes bytes
- Use `strings.Builder`, which doesn't copy to output string

> Thank you for contributing to Coraza WAF, your effort is greatly appreciated
> Before submitting check if what you want to add to `coraza` list meets [quality standards](https://github.com/corazawaf/coraza/blob/master/CONTRIBUTING.md#quality-standards) before sending pull request. Thanks!

**Note**: _that go.mod and go.sum can only be modified for tested dependency updates or justified new features._

**Make sure that you've checked the boxes below before you submit PR:**

- [X] My code includes positive and negative tests.
- [X] I have an appropriate description with correct grammar.
- [X] I have read [Contribution guidelines](https://github.com/corazawaf/coraza/blob/master/CONTRIBUTING.md#contribution-guidelines), [maintainers note](https://github.com/corazawaf/coraza/blob/master/CONTRIBUTING.md#maintainers) and [Quality standard](https://github.com/github.com/corazawaf/coraza/v2sso/coraza-waf/blob/master/CONTRIBUTING.md#quality-standards).

Thanks for your PR :heart:

After
```
goos: darwin
goarch: arm64
pkg: github.com/corazawaf/coraza/v3/transformations
BenchmarkB64Decode
BenchmarkB64Decode/VGVzdENhc2U=
BenchmarkB64Decode/VGVzdENhc2U=-10         	16657209	        68.79 ns/op
BenchmarkB64Decode/P.HNjcmlwdD5hbGVydCgxKTwvc2NyaXB0Pg==
BenchmarkB64Decode/P.HNjcmlwdD5hbGVydCgxKTwvc2NyaXB0Pg==-10         	183049256	         6.552 ns/op
BenchmarkB64Decode/VGVzdABDYXNl
BenchmarkB64Decode/VGVzdABDYXNl-10                                  	17273186	        68.68 ns/op
PASS
```


Before
```
goos: darwin
goarch: arm64
pkg: github.com/corazawaf/coraza/v3/transformations
BenchmarkB64Decode
BenchmarkB64Decode/VGVzdENhc2U=
BenchmarkB64Decode/VGVzdENhc2U=-10         	14609536	        79.49 ns/op
BenchmarkB64Decode/P.HNjcmlwdD5hbGVydCgxKTwvc2NyaXB0Pg==
BenchmarkB64Decode/P.HNjcmlwdD5hbGVydCgxKTwvc2NyaXB0Pg==-10         	32041297	        37.65 ns/op
BenchmarkB64Decode/VGVzdABDYXNl
BenchmarkB64Decode/VGVzdABDYXNl-10                                  	15384277	        76.53 ns/op
PASS
```